# frontend/src/types/openapi.yaml
...
openapi: 3.0.1
info:
  title: travelPlan API Test
  description: travelPlan의 api 테스트입니다.
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  "/place/{id}":
    get:
      tags:
      - Place API
      summary: 여행지 조회
      operationId: getPlace
      parameters:
      - name: id
        in: path
        description: 여행지 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataPlaceResDto"
    put:
      tags:
      - Place API
      summary: 여행지 정보 수정
      operationId: updatePlace
      parameters:
      - name: id
        in: path
        description: 여행지 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PlaceUpdateReqDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataPlaceResDto"
    delete:
      tags:
      - Place API
      summary: 여행지 삭제
      operationId: deletePlace
      parameters:
      - name: id
        in: path
        description: 여행지 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataVoid"
  "/member/{id}":
    put:
      tags:
      - Member API
      summary: 회원정보 수정
      operationId: updateMember
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MemberUpdateRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MemberResponseDto"
    delete:
      tags:
      - Member API
      summary: 회원 삭제
      operationId: deleteMember
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
  "/api/reviews/{reviewId}":
    get:
      tags:
      - review-controller
      operationId: getReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataReviewResponseDto"
    put:
      tags:
      - review-controller
      operationId: updateReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: memberId
        in: query
        required: false
        schema:
          type: integer
          format: int64
          default: 1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ReviewRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataReviewResponseDto"
    delete:
      tags:
      - review-controller
      operationId: deleteReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: memberId
        in: query
        required: false
        schema:
          type: integer
          format: int64
          default: 1
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataVoid"
  "/api/comments/{commentId}":
    get:
      tags:
      - comment-controller
      operationId: getComment
      parameters:
      - name: commentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataCommentResponseDto"
    put:
      tags:
      - comment-controller
      operationId: updateComment
      parameters:
      - name: commentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: memberId
        in: query
        required: false
        schema:
          type: integer
          format: int64
          default: 1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CommentRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataCommentResponseDto"
    delete:
      tags:
      - comment-controller
      operationId: deleteComment
      parameters:
      - name: commentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: memberId
        in: query
        required: false
        schema:
          type: integer
          format: int64
          default: 1
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataVoid"
...
  "/api/reviews":
    get:
      tags:
      - review-controller
      operationId: getReviews
      parameters:
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: newest
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: placeId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: memberId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListReviewResponseDto"
    post:
      tags:
      - review-controller
      operationId: createReview
      parameters:
      - name: memberId
        in: query
        required: false
        schema:
          type: integer
          format: int64
          default: 1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ReviewRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataReviewResponseDto"
  "/api/comments":
    post:
      tags:
      - comment-controller
      operationId: createComment
      parameters:
      - name: memberId
        in: query
        required: false
        schema:
          type: integer
          format: int64
          default: 1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CommentRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataCommentResponseDto"
...
  "/api/reviews/popular":
    get:
      tags:
      - review-controller
      operationId: getPopularReviews
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListReviewResponseDto"
  "/api/reviews/place/{placeId}":
    get:
      tags:
      - review-controller
      operationId: getReviewsByPlace
      parameters:
      - name: placeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListReviewResponseDto"
  "/api/comments/review/{reviewId}":
    get:
      tags:
      - comment-controller
      operationId: getCommentsByReview
      parameters:
      - name: reviewId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListCommentResponseDto"
  "/api/comments/member/{memberId}":
    get:
      tags:
      - comment-controller
      operationId: getCommentsByMember
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListCommentResponseDto"
...
components:
  schemas:
...
    PlaceUpdateReqDto:
      type: object
      properties:
        cityName:
          type: string
        placeName:
          type: string
        description:
          type: string
        category:
          type: string
          enum:
          - PLACE
          - STAY
          - RESTAURANT
          - CAFE
          - ETC
    PlaceResDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cityName:
          type: string
        placeName:
          type: string
        description:
          type: string
        category:
          type: string
          enum:
          - PLACE
          - STAY
          - RESTAURANT
          - CAFE
          - ETC
    RsDataPlaceResDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          "$ref": "#/components/schemas/PlaceResDto"
    MemberUpdateRequestDto:
      type: object
      properties:
        nickname:
          type: string
        email:
          type: string
        password:
          type: string
        profileImage:
          type: string
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        ageRange:
          type: string
          enum:
          - TEENS
          - TWENTIES
          - THIRTIES
          - FORTIES_PLUS
        travelStyle:
          type: string
          enum:
          - TOURISM
          - RELAXATION
          - SHOPPING
        aboutMe:
          type: string
    MemberResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        nickname:
          type: string
        profileImage:
          type: string
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        ageRange:
          type: string
          enum:
          - TEENS
          - TWENTIES
          - THIRTIES
          - FORTIES_PLUS
        travelStyle:
          type: string
          enum:
          - TOURISM
          - RELAXATION
          - SHOPPING
        aboutMe:
          type: string
        rating:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        authority:
          type: string
    ReviewRequestDto:
      required:
      - content
      - title
      type: object
      properties:
        title:
          maxLength: 30
          minLength: 2
          type: string
        content:
          maxLength: 2000
          minLength: 10
          type: string
        rating:
          maximum: 5
          minimum: 1
          type: number
          format: double
        placeId:
          type: integer
          format: int64
    ReviewResponseDto:
      type: object
      properties:
        reviewId:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        rating:
          type: number
          format: double
        memberId:
          type: integer
          format: int64
        memberName:
          type: string
        placeId:
          type: integer
          format: int64
        viewCount:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        commentCount:
          type: integer
          format: int32
    RsDataReviewResponseDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          "$ref": "#/components/schemas/ReviewResponseDto"
    CommentRequestDto:
      required:
      - content
      type: object
      properties:
        content:
          maxLength: 100
          minLength: 2
          type: string
        reviewId:
          type: integer
          format: int64
    CommentResponseDto:
      type: object
      properties:
        commentId:
          type: integer
          format: int64
        content:
          type: string
        reviewId:
          type: integer
          format: int64
        memberId:
          type: integer
          format: int64
        memberName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RsDataCommentResponseDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          "$ref": "#/components/schemas/CommentResponseDto"
...
    RsDataListReviewResponseDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          type: array
          items:
            "$ref": "#/components/schemas/ReviewResponseDto"
    RsDataListCommentResponseDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          type: array
          items:
            "$ref": "#/components/schemas/CommentResponseDto"
...
    RsDataVoid:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          type: object
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT